# Modules/Pos/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zeebroo POS API
  version: 2.0.0
  description: |
    Full REST API for the Zeebroo POS platform — catalog, checkout, sales, invoicing,
    quotations, purchasing, expenses, payroll, design studio, and more.

    ## Authentication
    1. `POST /auth/token` with `email` + `password` → receive `access_token`
    2. Send `Authorization: Bearer {access_token}` on every protected route.

    ## Business scope
    Most endpoints are scoped to a business. Send **one** of:
    - Header `X-Business-Id: {id}` *(recommended)*
    - Query param `?business_id={id}`

    If omitted the API falls back to the authenticated user's most-recently-selected business.

    ## Branch scope
    When branch-level POS separation is enabled, additionally send `X-Branch-Id: {id}`.

    ## Response envelope
    Successful responses follow `{ "data": … }`.  Lists add `{ "data": […], "meta": { pagination } }`.
    Errors return `{ "message": "…", "errors": { "field": ["…"] } }` with an appropriate HTTP status.
servers:
  - url: /api/v1/pos
    description: POS API (prepend your app URL, e.g. https://platform.zeebroo.com)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
...
paths:
...
  /hr/payroll/rule-sets/{ruleSet}:
    get:
      tags: [HR payroll rule sets]
      summary: Get rule set detail with rules
      parameters:
        - { name: ruleSet, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Rule set detail
    patch:
      tags: [HR payroll rule sets]
      summary: Update a rule set
      parameters:
        - { name: ruleSet, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:         { type: string }
                currency:     { type: string }
                effective_to: { type: string, format: date }
                is_default:   { type: boolean }
                is_active:    { type: boolean }
                notes:        { type: string }
      responses:
        '200':
          description: Updated
    delete:
      tags: [HR payroll rule sets]
      summary: Delete a rule set
      parameters:
        - { name: ruleSet, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Deleted
...
  /hr/payroll/rule-sets/{ruleSet}/rules/{rule}:
    patch:
      tags: [HR payroll rule sets]
      summary: Update a payroll rule
      parameters:
        - { name: ruleSet, in: path, required: true, schema: { type: integer } }
        - { name: rule,    in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:        { type: string }
                sort_order:  { type: integer }
                is_taxable:  { type: boolean }
                is_statutory: { type: boolean }
                is_active:   { type: boolean }
                config_json: { type: string }
      responses:
        '200':
          description: Updated
    delete:
      tags: [HR payroll rule sets]
      summary: Delete a payroll rule
      parameters:
        - { name: ruleSet, in: path, required: true, schema: { type: integer } }
        - { name: rule,    in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Deleted
...